Add skeletons, split pack display for passed cards and some optimizations (#22)
* First take on skeletons

* Improvement to skeletons

* Skeletons now display the state vs what is in the deck

* Clear skeleton state when needed

* Toggle showing of card counts in skeletons

* Annotations

* Intial take on annotations from the crypt

* Fix annotations

* Add a bunch of missing files

* Make sure we listen on an actual log

* Better skeleton annotation name

* Skeleton suggestions improvements

* Set filtering and app settings

* Missing files

* Preserve the order of cards in skeleton

* Allow empty ranges on cards

* More range tweaks

* Don't transform the names

* Imporvements to image rendering speed

* Ignore redundant properties in DraftState when serializing

* Fix format of the Arena DB changing

* Add the ability to split passed cards by pack and either stack them or not

+++ CardPile.CardData.SeventeenLands/RawWinData.cs
@@ -9,8 +9,8 @@ internal class RawWinData
         bool? isSummary,
         string? colorName,
         string? shortColorName,
-        int? wins,
-        int? games,
+        double? wins,
+        double? games,
         Color colors
     )
     {
@@ -28,8 +28,8 @@ internal class RawWinData
         bool? is_summary,
         string? color_name,
         string? short_name,
-        int? wins,
-        int? games
+        double? wins,
+        double? games
     ) : this(is_summary, color_name, short_name, wins, games, Utils.ParseColors(short_name))
     {}
 
@@ -39,9 +39,9 @@ internal class RawWinData
 
     internal string? ShortColorName { get; init; }
 
-    internal int? Wins { get; set; }
+    internal double? Wins { get; set; }
 
-    internal int? Games { get; init; }
+    internal double? Games { get; init; }
 
     internal Color Colors { get; init; }
 }
